TIKA-4720: merge and clean
