+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-integration-tests/src/test/resources/test-documents/testHTML_head.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<head>
+  <meta charset="utf-8">
+  <meta http-equiv="x-dns-prefetch-control" content="on">
+  <meta name="viewport" id="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
+  <script src="https://tika.apache.org/logic0.js" async defer></script>
+  <script src="https://tika.apache.org/logic1.js" defer></script>
+  <script src="https://tika.apache.org/logic2.js" defer></script>
+  <link rel="stylesheet" href="https://tika.apache.org/style0.css">
+  <link rel="stylesheet" href="https://tika.apache.org/style1.css">
+  <link rel="stylesheet" href="https://tika.apache.org/style2.css">
+  <link rel="stylesheet" href="https://tika.apache.org/style3.css">
+  <link rel="stylesheet" href="https://tika.apache.org/style4.css">
+  <link rel="canonical" href="https://tika.apache.org/">
+  <link rel="mask-icon" sizes="any" href="https://tika.apache.org/images/tika.svg" color="#aaaaaa">
+  <link rel="icon" href="https://tika.apache.org/favicon.png">
+  <link rel="manifest" href="/manifest.json">
+  <meta name="application-name" content="Tika">
+  <meta name="referrer" content="origin">
+  <meta property="fb:app_id" content="111111111111">
+  <meta property="og:site_name" content="Tika">
+  <script src="https://tika.apache.org/logic3.js" defer></script>
+  <script src="https://tika.apache.org/logic4.js" defer></script>
+  <script src="https://tika.apache.org/logic5.js" defer></script>
+  <title>Apache Tika</title>
+  <link rel="alternate" hreflang="en" href="https://tika.apache.org/" />
+  <link rel="alternate" href="ios-app://111111111/tika/home" />
+  <link rel="alternate" href="android-app://org.apache/tika/" />
+</head>
+<body>
+</body>
+</html>
